Add word list input as separate text files & enhanced game dev.
Add word list input as separate text files & a notebook to extended the enhanced  version of hangman.

To improve my hangman game, I decided to add a difficulty level selector, us a text file as the source data for my words list and to introduce the "classic hangman doddle image" as a representation of the players' lives.